Added migration of tabs for chrome & safari
A EventScripts script for mains and battery (power_management) and for
wake up and EventScripts start (wakes). These scripts only contain
minimal code, everything they call is in even_lib. The code which
deals with the tab migration between Chrom and Safari is in
chrom_to_safari; this is called by even_lib.