Bug fix for Chrome
If Chrome is running, without any browser window open but with a running
extension which has an open window, it would not create any tabs. I
guess because the extension don't support tab creation or more likely
they don't count as a window object. At least the don't count towards
the count of open window objects which is how the problem is solved.